Repo init
As usual with my projects, things become a mess
Nudging up to 500 loc the structure is a bit tenuous at best.

The idea here is to have a structure to quickly profile the performance
of buffering files into ram. Specifically, an alloc'd region of memory
and then reading via fread.

the slowload section does just that. The entire program waits whilst a
single thread loads the file into memory.

uringload utilizes the new io_uring kernel module to load the files in
asynchronously.

The goal of this exercise is to profile the loading ---AND--- a
subsequent arbitrary operation done on said loaded data.
Of course, the goal is to see if I can write a program that starts to
process incoming read data before the read operation is considered
complete. In theory, this should be faster.